+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fixed bug where [custom field](https://plugins.doublesecretagency.com/notifier/custom-fields) values were empty in outbound messages.
 - Fixed bug where [Dynamic Recipients](https://plugins.doublesecretagency.com/notifier/recipients/types/dynamic-recipients) couldn't accept a single element.
 
 ## 3.2.1 - 2026-08-01